Added new event: ClientCreated.
Added a new event that is thrown when the server creates a client. This
event MUST return a new TcpClientHandler (the client) when the method
returns.

Useful for when you want to use a customized client that extends from
TcpClientHandler.

If you don't want to use the event, then the client will automatically
create a client on it's own.